Moved the dropdown into its own directory. Moved dropdown content to its own component. Tweaked designed, reduced margin top to accomdoate when navbar shrinks when scrolling. Moved the button and subtotal info into the conditional.
