+++ src/containers/Header/Header.css
@@ -147,11 +147,15 @@ a:active {
 
 .cart-popover-menu {
     font-family: "Montserrat", sans-serif;
+    font-weight: 400;
     max-width: 20rem;
-    margin-top: 26px;
+    margin-top: 11px;
     padding: 20px;
-    border: 1px solid #ffffff;
-    box-shadow: 1px 1px 1px 1px rgba(0, 0, 0, 0.1);
+    border: none !important;
+    border-radius: 0;
+    box-shadow: 0 8px 15px rgba(0,0,0,.1);
+    transform: 0.3s;
+    opacity: 0.3s;
 }
 
 .cart-item-img {
@@ -167,6 +171,7 @@ a:active {
 .cart-item-title:hover {
   color: #66de80 !important;
   cursor: pointer !important;
+  transition: 0.5s;
 }
 
 .cart-item-close-action {
